requirements.txt
asgiref==3.8.1
black==24.4.2
boto3==1.34.117
botocore==1.34.117
click==8.1.7
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
django-storages==1.14.3
djangorestframework==3.15.1
Faker==25.3.0
iniconfig==2.0.0
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-django==4.8.0
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
